+++ ros/drake_cars_examples/src/car_sim_lcm_and_ros.cc
@@ -28,6 +28,7 @@ namespace {
 using drake::examples::cars::CreateRigidBodySystem;
 using drake::examples::cars::CreateVehicleSystem;
 using drake::examples::cars::GetCarSimulationDefaultOptions;
+using drake::examples::cars::ParseDuration;
 
 /** Driving Simulator
  * Usage:  car_sim_lcm_and_ros vehicle_model_file [world_model files ...]
@@ -38,12 +39,9 @@ int do_main(int argc, const char* argv[]) {
   // Initializes the communication layer.
   std::shared_ptr<lcm::LCM> lcm = std::make_shared<lcm::LCM>();
 
-  // Instantiates a duration variable that will be set by the call to
-  // CreateRigidBodySystem() below.
-  double duration = std::numeric_limits<double>::infinity();
-
   // Initializes the rigid body system.
-  auto rigid_body_sys = CreateRigidBodySystem(argc, argv, &duration);
+  auto rigid_body_sys = CreateRigidBodySystem(argc, argv);
+
   auto const& tree = rigid_body_sys->getRigidBodyTree();
 
   // Initializes and cascades all of the other systems.
@@ -92,6 +90,10 @@ int do_main(int argc, const char* argv[]) {
   // Defines the start time of the simulation.
   const double kStartTime = 0;
 
+  // Initializes and obtains the desired simulation duration.
+  double duration = ParseDuration(argc, argv);
+
+  // Starts the simulation.
   drake::ros::run_ros_vehicle_sim(sys, kStartTime, duration, x0, options);
 
   return 0;