+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Unreleased: changes on master, not yet released
 [//]: # "Altered functionality or APIs."
 ### Changed
 
+ - [#3014][] Removed duration parameter from CreateRigidBodySystem() in car_simulation.h
  - [#3003][] Made `RigidBodyFrame` member variables private. Added accessors.
  - [#2997][] Renamed drake/Path.h to drake/common/drake_path.h
  - [#2983][] Renamed namespace `Drake` to be `drake`.
@@ -106,3 +107,4 @@ Changes in version v0.9.11 and before are not provided.
 [#2983]: https://github.com/RobotLocomotion/drake/issues/2983
 [#2997]: https://github.com/RobotLocomotion/drake/issues/2997
 [#3003]: https://github.com/RobotLocomotion/drake/issues/3003
+[#3014]: https://github.com/RobotLocomotion/drake/issues/3014